refactor BotCore
