+++ src/main/java/com/upravad/cookbot/core/options/Options.java
@@ -1,4 +1,4 @@
-package com.upravad.cookbot.core;
+package com.upravad.cookbot.core.options;
 
 public interface Options {
