+++ src/main/java/com/upravad/cookbot/view/ButtonsView.java
@@ -1,11 +1,11 @@
 package com.upravad.cookbot.view;
 
-import static com.upravad.cookbot.util.Messages.ORDER;
 import static com.upravad.cookbot.util.Messages.SBER_URL;
+import static com.upravad.cookbot.util.Messages.ORDER;
 
-import com.upravad.cookbot.database.dto.IngredientDto;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
+import com.upravad.cookbot.database.dto.IngredientDto;
 import org.springframework.stereotype.Component;
 import com.upravad.cookbot.database.dto.DishDto;
 import java.util.ArrayList;
@@ -17,8 +17,8 @@ public class ButtonsView implements View {
   /**
    * Category options.
    *
-   * @param dishes
-   * @return
+   * @param dishes from service
+   * @return prepared InlineKeyboardMarkup for category buttons
    */
   public InlineKeyboardMarkup getDishesButtons(List<DishDto> dishes) {
     var markup = new InlineKeyboardMarkup();
@@ -39,10 +39,12 @@ public class ButtonsView implements View {
         .build());
   }
 
+
   /**
    * Recipe options.
    *
-   * @return
+   * @param dish from CallbackQuery
+   * @return prepared InlineKeyboardMarkup for recipe buttons
    */
   public InlineKeyboardMarkup getPhotoButtons(DishDto dish) {
     var markup = new InlineKeyboardMarkup();
@@ -52,17 +54,17 @@ public class ButtonsView implements View {
 
   private List<List<InlineKeyboardButton>> getPhotoRows(DishDto dish) {
     List<List<InlineKeyboardButton>> rows = new ArrayList<>();
-    rows.add(getRow());
+    rows.add(getHeaderRow("Ингредиенты: ", "ingredients"));
     dish.getIngredients().forEach(ingredient -> rows.add(getIngredientRow(ingredient)));
     rows.add(getCostRow(dish));
     rows.add(getOrderRow());
     return rows;
   }
 
-  private List<InlineKeyboardButton> getRow() {
+  private List<InlineKeyboardButton> getHeaderRow(String text, String data) {
     return List.of(InlineKeyboardButton.builder()
-        .text("Ингредиенты: ")
-        .callbackData("ingredients")
+        .text(text)
+        .callbackData(data)
         .build());
   }
